start breaking down addBook functionalities
