+++ components/Modal/styles/modalStyles.ts
@@ -1,0 +1,13 @@
+import React from 'react';
+
+export const modalDiv: React.CSSProperties = {
+  position: 'absolute',
+  top: '50%',
+  left: '50%',
+  transform: 'translate(-50%, -50%)',
+  backgroundColor: 'white',
+  padding: '25px 40px',
+  borderRadius: '4px',
+  width: '80%',
+  maxWidth: '400px',
+};
