+++ components/bookCatalog/styles/addBookStyles.ts
@@ -1,0 +1,33 @@
+export const addBookStyle = {
+  color: 'black',
+  border: '1px solid black',
+  '&:hover': {
+    border: '1px solid black',
+  },
+};
+
+export const searchBookCatalogButton = {
+  margin: '1rem 0',
+  padding: '0',
+  color: 'black',
+  textDecoration: 'underline',
+};
+
+export const showSearchCatalogStyle = {
+  padding: '1rem',
+  borderRadius: '0.4rem',
+  backgroundColor: '#FFCC00',
+  '& p': {
+    margin: '0.5rem 0',
+  },
+  '& .MuiTextField-root': { m: 1, width: '25ch' },
+};
+
+export const searchBookCatalogStyle = {
+  color: 'black',
+  border: '1px solid black',
+  margin: '0.5rem 0',
+  '&:hover': {
+    border: '1px solid black',
+  },
+};
